Revert "Merge branch 'master' into interoper"
This reverts commit 6fdcd42d578c8e66bba47fea27bd259c69e94ac0, reversing
changes made to 5075453d8b458ec765f3852cf283c44db5b62789.

+++ libs/build.sh
@@ -1,11 +1,17 @@
 #! /bin/bash
 
+# Check arguments...
+if [ $# -ne 1 ] ; then
+    echo "usage: $0 <nc2|nc4>"
+    exit
+fi
+
 # Setup...
 target=$(rm -rf build && mkdir -p build && cd build && pwd)
 threads=$(cat /proc/cpuinfo | grep processor | wc -l)
 
 # Prepare directories...
-mkdir -p $target/src $target/bin $target/include $target/lib $target/man/man1 \
+mkdir -p $target/src $target/bin $target/lib $target/man/man1 \
     && cp *tar.bz2 $target/src \
     && cd $target/src \
     && for f in $(ls *tar.bz2) ; do tar xvjf $f ; done \
@@ -18,42 +24,28 @@ cd $target/src/$dir \
     && make -j$threads && make check && make install && make clean \
 	|| exit
 
-# zlib...
-dir=zlib-1.2.12
-cd $target/src/$dir \
-    && ./configure --prefix=$target \
-    && make -j$threads && make check && make install && make clean \
-	|| exit
-
-# zfp...
-dir=zfp-0.5.5
-cd $target/src/$dir \
-    && make && make test \
-    && cp -a bin/* $target/bin/ \
-    && cp -a include/* $target/include/ \
-    && cp -a lib/* $target/lib/ \
-	|| exit
-
-# zstd...
-dir=zstd-1.5.2
-cd $target/src/$dir \
-    && make -j$threads && make check \
-    && cp -a lib/libzstd* $target/lib/ \
-    && cp -a lib/*.h $target/include/ \
-    && make clean \
-	|| exit
-
-# HDF5...
-dir=hdf5-1.12.1
-cd $target/src/$dir \
-    && ./configure --prefix=$target --with-zlib=$target --enable-hl \
-    && make -j$threads ; make -j$threads && make check && make install && make clean \
-	|| exit
+# Check netCDF flag...
+if [ $1 = "nc4" ] ; then
+    
+    # zlib...
+    dir=zlib-1.2.11
+    cd $target/src/$dir \
+	&& ./configure --prefix=$target \
+	&& make -j$threads && make check && make install && make clean \
+	    || exit
+    
+    # HDF5...
+    dir=hdf5-1.12.1
+    cd $target/src/$dir \
+	&& ./configure --prefix=$target --with-zlib=$target --enable-hl \
+	&& make -j$threads ; make -j$threads && make check && make install && make clean \
+	    || exit
+fi
 
 # netCDF...
 dir=netcdf-c-4.8.1
 cd $target/src/$dir \
-    && CPPFLAGS=-I$target/include LDFLAGS=-L$target/lib ./configure --prefix=$target --disable-dap --disable-nczarr \
+    && CPPFLAGS=-I$target/include LDFLAGS=-L$target/lib ./configure --prefix=$target --disable-dap --disable-nczarr $([ $1 != "nc4" ] && echo "--disable-netcdf-4") \
     && make -j$threads && make check && make install && make clean \
 	|| exit
